Add file uploading and manifest generation
This commit introduces a new module for file uploading and manifest generating functionality. Now, the code walks through a directory, uploads the files found, and creates a manifest.json for tracking. Currently, the uploader targets the Bunny CDN, but commented code for Azure Storage Blobs is also included for future reference.